# Conflicts:
#	src/main/java/aseca/roobinhood/api/service/AuthenticationService.java
#	src/main/java/aseca/roobinhood/api/service/TransactionService.java
#	src/main/resources/application.yml

+++ src/main/java/aseca/roobinhood/api/dto/TransactionDto.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TransactionDto {
     private TickerDto tickerDto;
-    private Long amount;
-    private Long price;
+    private double amount;
+    private double price;
 }
